fix(catalog-glue): merge TestGlueCatalogSchema and TestGlueCatalogTable into TestGlueCatalogOperations
